Use a new name and avatar
This name is only used when posting via the REST API, the rest of the responses here go back via the bot API.
I've updated the botname to pull the bot name from the settings.  It's unclear whether you can access the bots avatar that way, so I haven't for now.