Add support for NaturalEngland as per PR#3
This code adds support for a list of approved address regular
expressions, and then loops through them.

It uses a slightly overclever "reduce", on the basis that any logical
equivalancy using OR means that any single True reduces to True, and all
False's reduced to False.

This means you only have to pass any one of the regular expressions, not
all of them, and you only fail if you don't meet any of the expressions.

Note, this adds NaturalEngland, separately we need to work out what
domains we want to add#3

This code adds support for a list of approved address regular
expressions, and then loops through them.

It uses a slightly overclever "reduce", on the basis that any logical
equivalancy using OR means that any single True reduces to True, and all
False's reduced to False.

This means you only have to pass any one of the regular expressions, not
all of them, and you only fail if you don't meet any of the expressions.

Note, this adds NaturalEngland, separately we need to work out what
domains we want to add#3

This code adds support for a list of approved address regular
expressions, and then loops through them.

It uses a slightly overclever "reduce", on the basis that any logical
equivalancy using OR means that any single True reduces to True, and all
False's reduced to False.

This means you only have to pass any one of the regular expressions, not
all of them, and you only fail if you don't meet any of the expressions.

Note, this adds NaturalEngland, separately we need to work out what
domains we want to add