autoresponse about the 'invite' functionality
Trying to get the bot to autorespond with instructions for inviting new people to the Slack channel. * , * 

I am not sure if removing ''direct_message,direct_mention,mention'; will have the intended effect of making it listen to all messages? Should it listen to messages only in the @chat channel? Is the autoresponse clear enough? Should it direct people right to DMs? Should the list of domains be elsewhere or is it clear enough? DOES IT WORK? So many questions I'm hoping someone who can computer better than me will answer. u__u