Ensure only one instance of the bot is running
Each instance now generates itself a UUID when it starts. We use a new
table in postgres, which the process puts this UUID in. When it notices
that the UUID in the table no longer matches this value, the process
exits.

Transactions, and the NOTIFY and LISTEN features in postgres are used to
ensure that clients notice when a new value has been set in the
database.

The overall effect should be that the most recently started instance
will remain running, and all other instances would stop.

I believe this will fix issue #30.  However, this somewhat depends on
the behaviour of CloudFoundry. If CloudFoundry attempts to continually
restart a second node, it will result in the application being
repeatedly stopped and started. I'm not sure how to test this other than
trying it.