Shut the process down on connection close
The slack bot previously had logic to retry starting the RTM connection
to Slack each time an error happened when starting it, or when an
rtm_close() message was received.  We've been seeing instances where the
slack bot sends multiple replies to messages, which I hypothesise is due
to multiple connections ending up open (perhaps the 'rtm_close' event
can be sent for a non-permanent close?  Or perhaps an error can be
returned by startRTM but the connection still be formed?)

This patch attempts to address that by shutting down the whole process
on rtm_close(), or on an error with startRTM.  In the case of an error,
a delay of 60 seconds is added to avoid a busy loop if slack is down for
a period of time.