Add CS Jobs URL shortener
The URLs usually shared in Slack from the Civil Service Jobs site are
usually unwieldy blobs of Base64 encoded parameters that are not nice to
share.

However, the short URL is often available on the page that gets shared,
albeit hidden until one clicks a link to share on social media.

This commit adds the feature to xgovslackbot to listen to those long
URLs (in the channels it's invited to) and shorten them.

If the user explicitly asks xgovslackbot to shorten it, the short URL
will be public. If the user doesn't, the bot will quietly message them
using Slack's ephemeral feature (like when you upload a file beyond
Slack's size limit) -- so that they can edit their message to shorten
the URL if they choose to do so.